requirements.txt
torch
sentence-transformers
langchain
langchain-community
transformers
pdf2image
pillow
opencv-python
pytesseract
pdfplumber
pypdf
pandas
gradio
faiss-cpu
python-dotenv
tabulate
sentencepiece

